requirements.txt
absl-py==2.4.0
    # via
    #   keras
    #   tensorboard
    #   tensorflow
albucore==0.0.24
    # via albumentations
albumentations==2.0.8
    # via -r base.in
altair==6.0.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
astunparse==1.6.3
    # via tensorflow
attrs==25.4.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via
    #   flask
    #   streamlit
cachetools==6.2.6
    # via streamlit
certifi==2026.1.4
    # via requests
charset-normalizer==3.4.4
    # via requests
click==8.3.1
    # via
    #   flask
    #   streamlit
colorama==0.4.6
    # via click
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
defusedxml==0.7.1
    # via fpdf2
flask==3.1.0
    # via -r base.in
flatbuffers==25.12.19
    # via tensorflow
fonttools==4.61.1
    # via
    #   fpdf2
    #   matplotlib
fpdf2==2.7.6
    # via -r base.in
gast==0.7.0
    # via tensorflow
gitdb==4.0.12
    # via gitpython
gitpython==3.1.46
    # via streamlit
google-pasta==0.2.0
    # via tensorflow
grpcio==1.78.0
    # via
    #   tensorboard
    #   tensorflow
h5py==3.15.1
    # via
    #   keras
    #   tensorflow
idna==3.11
    # via requests
imageio==2.37.2
    # via scikit-image
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via
    #   altair
    #   flask
    #   pydeck
jsonschema==4.26.0
    # via altair
jsonschema-specifications==2025.9.1
    # via jsonschema
keras==3.13.2
    # via tensorflow
kiwisolver==1.4.9
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
libclang==18.1.1
    # via tensorflow
markdown==3.10.2
    # via tensorboard
markdown-it-py==4.0.0
    # via rich
markupsafe==3.0.3
    # via
    #   jinja2
    #   werkzeug
matplotlib==3.10.8
    # via
    #   -r base.in
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
ml-dtypes==0.5.4
    # via
    #   keras
    #   tensorflow
namex==0.1.0
    # via keras
narwhals==2.16.0
    # via altair
networkx==3.6.1
    # via scikit-image
numpy==2.4.2
    # via
    #   -r base.in
    #   albucore
    #   albumentations
    #   contourpy
    #   h5py
    #   imageio
    #   keras
    #   matplotlib
    #   ml-dtypes
    #   opencv-python-headless
    #   pandas
    #   pydeck
    #   scikit-image
    #   scipy
    #   seaborn
    #   streamlit
    #   tensorboard
    #   tensorflow
    #   tifffile
opencv-python-headless==4.13.0.92
    # via
    #   -r base.in
    #   albucore
    #   albumentations
opt-einsum==3.4.0
    # via tensorflow
optree==0.18.0
    # via keras
packaging==26.0
    # via
    #   altair
    #   keras
    #   lazy-loader
    #   matplotlib
    #   scikit-image
    #   streamlit
    #   tensorboard
    #   tensorflow
    #   wheel
pandas==2.3.3
    # via
    #   -r base.in
    #   seaborn
    #   streamlit
pillow==12.1.1
    # via
    #   fpdf2
    #   imageio
    #   matplotlib
    #   scikit-image
    #   streamlit
    #   tensorboard
prometheus-client==0.21.1
    # via -r base.in
protobuf==6.33.5
    # via
    #   streamlit
    #   tensorboard
    #   tensorflow
psutil==6.1.1
    # via -r base.in
pyarrow==23.0.0
    # via streamlit
pydantic==2.12.5
    # via albumentations
pydantic-core==2.41.5
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via rich
pyparsing==3.3.2
    # via matplotlib
python-dateutil==2.9.0.post0
    # via
    #   matplotlib
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via albumentations
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.5
    # via
    #   -r base.in
    #   streamlit
    #   tensorflow
rich==14.3.2
    # via keras
rpds-py==0.30.0
    # via
    #   jsonschema
    #   referencing
scikit-image==0.26.0
    # via -r base.in
scipy==1.17.0
    # via
    #   -r base.in
    #   albumentations
    #   scikit-image
seaborn==0.13.2
    # via -r base.in
sentry-sdk==2.19.2
    # via -r base.in
simsimd==6.5.12
    # via albucore
six==1.17.0
    # via
    #   astunparse
    #   google-pasta
    #   python-dateutil
    #   tensorflow
smmap==5.0.2
    # via gitdb
streamlit==1.54.0
    # via -r base.in
stringzilla==4.6.0
    # via albucore
tenacity==9.1.4
    # via streamlit
tensorboard==2.20.0
    # via tensorflow
tensorboard-data-server==0.7.2
    # via tensorboard
tensorflow==2.20.0
    # via -r base.in
termcolor==3.3.0
    # via tensorflow
tifffile==2026.1.28
    # via scikit-image
toml==0.10.2
    # via streamlit
tornado==6.5.4
    # via streamlit
typing-extensions==4.15.0
    # via
    #   altair
    #   grpcio
    #   optree
    #   pydantic
    #   pydantic-core
    #   referencing
    #   streamlit
    #   tensorflow
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.3
    # via pandas
urllib3==2.6.3
    # via requests
watchdog==6.0.0
    # via streamlit
werkzeug==3.1.5
    # via
    #   flask
    #   tensorboard
wheel==0.46.3
    # via astunparse
wrapt==2.1.1
    # via tensorflow

# The following packages are considered to be unsafe in a requirements file:
# setuptools
